refactor: improve code layering of ozhera-prometheus-agent module
