+++ ozhera-prometheus-agent/ozhera-prometheus-agent-service/src/main/java/com/xiaomi/youpin/prometheus/agent/service/api/MioneMachineServiceExtension.java
@@ -13,6 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
+
 package com.xiaomi.youpin.prometheus.agent.service.api;
 
 import com.xiaomi.youpin.prometheus.agent.domain.Ips;
@@ -24,5 +25,6 @@ import java.util.List;
  */
 
 public interface MioneMachineServiceExtension {
+    
     List<Ips> queryMachineList(String type);
 }
